move checkPersistentDisks() to DiskAdminService
